refactor(html): move tailwind gen
